refactor(tests): rename NamingConventions namespace to NamingOverrideTable
- Updated namespace across integration test classes and infrastructure to reflect the new name.
- Adjusted type definitions, imports, and references to align with the updated namespace.
- Improved clarity and organization of test files by restructuring namespaces.
